# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "{JSON} Placeholder API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/comments/{id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Comment By ID",
                "operationId": "comments.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "postId": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "id labore ex et quam laborum"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "Eliseo@gardner.biz"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "laudantium enim quasi est quidem magnam voluptate ipsam eos\\ntempora quo necessitatibus\\ndolor quam autem quasi\\nreiciendis et nam sapiente accusantium"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "comments"
                ],
                "summary": "Delete existing Comment",
                "operationId": "comments.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully delete data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "comments"
                ],
                "summary": "Update comment",
                "operationId": "comments.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payload for update existing Comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "body",
                                    "userId"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "foo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully update data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            }
        },
...
        "/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Post By ID",
                "operationId": "posts.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit"
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "posts"
                ],
                "summary": "Update existing Post",
                "operationId": "posts.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payload for update existing Post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body",
                                    "userId"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "foo"
                                    },
                                    "body": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bar"
                                    },
                                    "userId": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully update data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete existing Post",
                "operationId": "posts.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully delete data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "posts"
                ],
                "summary": "Update existing Post",
                "operationId": "posts.updatePatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Payload for update existing Post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "body",
                                    "userId"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "foo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully update data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "mobilekey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    }
}